fix MSVC warning about comparison signed <-> unsigned

+++ src/gui/dialogs/advanced_graphics_options.cpp
@@ -104,7 +104,7 @@ void tadvanced_graphics_options::scale_button_callback(std::string pref_id, SCAL
 
 	for (size_t x = 0; x < SCALING_ALGORITHM_COUNT; ++x) {
 		ttoggle_button * b = &find_widget<ttoggle_button>(&window, pref_id + "_" + SCALING_ALGORITHM_to_string(static_cast<SCALING_ALGORITHM>(x)), false);
-		b->set_value(x == me);
+		b->set_value(x == static_cast<size_t>(me));
 	}
 
 	image::update_from_preferences();
